yolov7_gui_standalone/requirements.txt
# YOLOv7 GUI 기본 요구사항
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
matplotlib>=3.3.0
numpy>=1.21.0
PyYAML>=5.4.0
tqdm>=4.60.0
Pillow>=8.3.0
pandas>=1.3.0
